requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
cairocffi==0.6
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0.3
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.7
django-appconf==1.0.4
django-cors-headers==3.5.0
django-imagekit==4.0.2
django-import-export==2.4.0
django-jazzmin==2.2.9
django-safedelete==0.5.6
django-softdelete==0.9.1
django-taggit==1.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-generators==0.5.0
et-xmlfile==1.0.1
Faker==4.1.3
fpdf==1.7.2
future==0.18.2
html5lib==1.1
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.5
pilkit==2.0
Pillow==8.4.0
pip-tools==4.4.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.53
requests==2.24.0
six==1.14.0
soft-delete==0.2.2
sqlparse==0.3.0
tablib==2.0.0
text-unidecode==1.3
tinycss2==1.0.2
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
ffmpeg-python==0.2.0
gunicorn
psycopg2-binary
SpeechRecognition
moviepy
icecream
